refactor: move sulfuras logic
Since there are several nested if statements specifying action to take
if item name is NOT Sulfuras, this simplifies the code by removing all
statements stating `if (items[i].name != 'Sulfuras, Hand of Ragnaros')`
and instead creating a case for `if (items[i].name === 'Sulfuras...')`
at the top. Used `break` because Sulfuras items do not change in sell_in
or quality, though perhaps this could be changed later.

Note that none of the statements under each `if (items[i].name !=
'Sulfuras, Hand of Ragnaros')` were removed, just the if statements
themselves.