Refactor: extract functions to mutate quality/sellin
Move all logic that mutates quality/sellin to separate functions so that it can be adjusted later when we want to return a fresh set of data.
Also handle special rules pertaining to quality limits on items (cannot be over 50) within the extracted functions.

Extract these functions, next plan to consolidate the numbers of times these functions are called.
This will hopefully make it easier to return a fresh set of data later on.