Refactor: reduce number of quality mutations
Do we still have if statements nested within if statements? Yes... and this logic could arguably be neatened up in a future refactoring.
This refactoring reduces the number of mutations to items. For instance, with backstage passes, instead of incrementing quality once, then again if there are 10 sell in days or less, then again if there are 5 sell in days or less, logic has been updated so that the quality is only modified once for each item.
This (I hope?) makes it smoother and less redundant when it comes to creating a fresh set of data rather than mutating the data later on.