Feat: Add support for printing items in an ordered list
Created a new file (item_list.js), imported the Item function from gilded_rose.js,
and added the list of commented out items from gilded_rose.js using the Item constructor
function.
Created a function in item_list.js that accepts an array as an argument and writes each item inside
of the array as a new list item in an HTML ordered list. The function then returns
the HTML list. To display object data properly on the Express server, each list item is wrapped
inside of the JSON.stringify() method.
Exported the function for getting an HTML list from an array, imported said
function into the test file, and then created a new series of tests, including tests for an array
with one item, an array with two items, an empty array, and an array of one item from the
gilded rose store.
Then, passed gilded rose items into the getHtmlListFromArray function, saved the resulting list
as a new variable, and sent this variable to display on the Express server.

FSA2021-38